* sni:
  Updating supervisor scripts with new virtualenv path.
  Adding virtualenvwrapper.
  Adding training Q to FAQ.
  Relative URI handling in feedparser changes when it doesn't have the URL of the feed. Thank goodness I caught this early, otherwise a bunch of sites' images would have been mysteriously broken.
  App role in upgrading virtualenv.
  Handling swap space correctly when upgrading pip requirements.
  Adding swapfile for installing pip requirements.
  Adding wsgi server to use new virtualenv gunicorn.
  Fixing issue with missing feed urls on original feed pages.
  Fixing feedfinder beautifulsoup issues, unicode issues in feed, and unicode comparisons issues with page_data (since it wasnt being decompressed when compared).
  Fix to scipy and the icon importer credited entirely to the work of @sbilly.
  Handling logging with virtualenv.
  -pdb
  Switching fabfile to virtualenv.
  mysql-python
  Attempting to fix SNI issues by using new feedfinder and using requests instead of feedparser. This needs virtualenv, stat.

+++ apps/rss_feeds/icon_importer.py
@@ -318,7 +318,8 @@ class IconImporter(object):
         # Reshape array of values to merge color bands. [[R], [G], [B], [A]] => [R, G, B, A]
         if len(shape) > 2:
             ar = ar.reshape(scipy.product(shape[:2]), shape[2])
-
+            ar = ar.astype(numpy.float)
+            
         # Get NUM_CLUSTERS worth of centroids.
         codes, _ = scipy.cluster.vq.kmeans(ar, NUM_CLUSTERS)
 
@@ -347,7 +348,7 @@ class IconImporter(object):
 
         # Find the most frequent color, based on the counts.
         index_max = scipy.argmax(counts)
-        peak = codes[index_max]
+        peak = codes.astype(int)[index_max]
         color = ''.join(chr(c) for c in peak).encode('hex')
 
         return color[:6]
